+++ src/components/Accounts/ChargeFeeFine/ItemInfo.js
@@ -54,6 +54,7 @@ class ItemInfo extends React.Component {
               {placeholder => (
                 <TextField
                   placeholder={placeholder}
+                  disabled={!this.props.editable}
                   onChange={this.onChangeSelectItem}
                 />
               )}