+++ src/components/Accounts/Menu.js
@@ -16,19 +16,20 @@ const Menu = (props) => {
   const outstanding = parseFloat(balance).toFixed(2);
   const showSelected = (selected !== 0 && selected !== parseFloat(0).toFixed(2))
     && outstanding > parseFloat(0).toFixed(2);
+  const buttonDisabled = !props.stripes.hasPerm('ui-users.feesfines.actions.all');
 
-  let type = 'Open';
+  let type = <FormattedMessage id="ui-users.accounts.open" />;
   if (query.layer === 'closed-accounts') {
-    type = 'Closed';
+    type = <FormattedMessage id="ui-users.accounts.closed" />;
   } else if (query.layer === 'all-accounts') {
-    type = 'All';
+    type = <FormattedMessage id="ui-users.accounts.all" />;
   }
 
   const firstMenu = (
     <div>
       <Row>
-        <Col>
-          <b>{`${type} fees/fines for `}</b>
+        <Col style={{ marginLeft: '20px' }}>
+          <b><FormattedMessage id="ui-users.accounts.history.statusLabel" values={{ type }} /></b>
           <Link to={`/users/view/${user.id}`}>
             {`${getFullName(user)} (${_.upperFirst(props.patronGroup.group)})`}
           </Link>
@@ -39,7 +40,7 @@ const Menu = (props) => {
         </Col>
       </Row>
       <Row>
-        <Col>
+        <Col style={{ marginLeft: '20px' }}>
           <FormattedMessage
             id="ui-users.accounts.outstanding"
             values={{
@@ -62,13 +63,13 @@ const Menu = (props) => {
 
   const lastMenu = (
     <div>
-      <Button onClick={e => props.onClickViewChargeFeeFine(e, {})}>
+      <Button marginBottom0 disabled={buttonDisabled} onClick={e => props.onClickViewChargeFeeFine(e, {})}>
         <FormattedMessage id="ui-users.accounts.button.new" />
       </Button>
-      <Button disabled={!actions.regularpayment} buttonStyle="primary" onClick={() => { props.onChangeActions({ regular: true }); }}><FormattedMessage id="ui-users.accounts.history.button.pay" /></Button>
-      <Button disabled={!actions.waive} buttonStyle="primary" onClick={() => { props.onChangeActions({ waiveMany: true }); }}><FormattedMessage id="ui-users.accounts.history.button.waive" /></Button>
-      <Button disabled buttonStyle="primary"><FormattedMessage id="ui-users.accounts.history.button.refund" /></Button>
-      <Button disabled buttonStyle="primary"><FormattedMessage id="ui-users.accounts.history.button.transfer" /></Button>
+      <Button marginBottom0 disabled={!((actions.regularpayment === true) && (buttonDisabled === false))} buttonStyle="primary" onClick={() => { props.onChangeActions({ regular: true }); }}><FormattedMessage id="ui-users.accounts.history.button.pay" /></Button>
+      <Button marginBottom0 disabled={!((actions.waive === true) && (buttonDisabled === false))} buttonStyle="primary" onClick={() => { props.onChangeActions({ waiveMany: true }); }}><FormattedMessage id="ui-users.accounts.history.button.waive" /></Button>
+      <Button marginBottom0 disabled buttonStyle="primary"><FormattedMessage id="ui-users.accounts.history.button.refund" /></Button>
+      <Button marginBottom0 disabled buttonStyle="primary"><FormattedMessage id="ui-users.accounts.history.button.transfer" /></Button>
     </div>);
 
   return (
@@ -78,6 +79,9 @@ const Menu = (props) => {
 
 Menu.propTypes = {
   user: PropTypes.object,
+  stripes: PropTypes.shape({
+    hasPerm: PropTypes.func,
+  }),
   showFilters: PropTypes.bool,
   balance: PropTypes.number,
   selected: PropTypes.number,